utils: Move download to utils
Removed unneeded package, containing only single util func.
