--- pkg/download/go.test
@@ -1,3 +1,0 @@
-#!/bin/sh
-
-go test -v .
